[Pytorch Edge] Support profiling kineto events from external source
Summary:
This diff exposes a way to add events to kineto profiler from external
source.
This can be a backend that executes a subgraph and wants to record this
execution in kineto profiler.
This diff also adds "backend" metadata to identify the backend an event
would have executed on.

Test Plan:
test_lite_interpreter

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: c6e93591ce8810e1d79df3f32ba407911e8f0629
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64397