OpInfos for some Tensor dtype conversion methods
OpInfos for:
- Tensor.bfloat16, Tensor.bool, Tensor.bypte, Tensor.char
- Tensor.double, Tensor.float, Tensor.half, Tensor.int
- Tensor.short, Tensor.long

None of these are supported by TorchScript. Also, the OpInfo autograd
test runner assumes that the operation is not allowed to change the
dtype of the argument, so only Tensor.double has
`supports_autograd=True` (in theory Tensor.bfloat16, Tensor.float,
Tensor.half should be differentiable).

Test Plan:
- run tests

ghstack-source-id: 9a23f6e9b4e30424542c033b36e4f541c7f7b2e7
Pull Request resolved: https://github.com/pytorch/pytorch/pull/64282